Per Ed Kmett's suggestion, trying to solve my problem of over-rigid type variables by making the types of stored functions existential. (Is that the right word?)
Impredicative types in GHC seem to be a problem, but wrapping them in
newtypes seems to alleviate it.

Now, among other issues, I am stumped by another problem: How do I
bring a type bound in a newtype definition into scope in the
expression I am feeding to that newtype's constructor?  (If that's
even what I want to do).